fix circular dep?
