+++ packages/notte-sdk/src/notte_sdk/endpoints/vaults.py
@@ -1,10 +1,15 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import TYPE_CHECKING, Unpack
+from typing import Unpack, final
 
+from notte_core.credentials.base import (
+    BaseVault,
+    CredentialsDict,
+    CreditCardDict,
+)
 from pydantic import BaseModel
-from typing_extensions import final, override
+from typing_extensions import override
 
 from notte_sdk.endpoints.base import BaseClient, NotteEndpoint
 from notte_sdk.types import (
@@ -14,6 +19,7 @@ from notte_sdk.types import (
     AddCreditCardRequest,
     AddCreditCardRequestDict,
     AddCreditCardResponse,
+    Credential,
     DeleteCredentialsRequest,
     DeleteCredentialsRequestDict,
     DeleteCredentialsResponse,
@@ -40,8 +46,48 @@ from notte_sdk.types import (
     VaultCreateResponse,
 )
 
-if TYPE_CHECKING:
-    from notte_sdk.vault import NotteVault
+
+# DEFINED HERE TO SIMPLIFY CIRCULAR DEPENDENCY
+# SHOULD ONLY BE INVOKED FROM ENDPOINT ANYWAY
+@final
+class NotteVault(BaseVault):
+    """Vault that fetches credentials stored using the sdk"""
+
+    def __init__(self, vault_id: str, vault_client: VaultsClient | None = None):
+        self.vault_id: str = vault_id
+
+        if vault_client is None:
+            vault_client = VaultsClient()
+
+        self.vault_client = vault_client
+
+    @override
+    def _add_credentials(self, url: str, creds: CredentialsDict) -> None:
+        _ = self.vault_client.add_or_update_credentials(self.vault_id, url=url, **creds)
+
+    @override
+    def _get_credentials_impl(self, url: str) -> CredentialsDict | None:
+        return self.vault_client.get_credentials(vault_id=self.vault_id, url=url).credentials
+
+    @override
+    def delete_credentials(self, url: str) -> None:
+        _ = self.vault_client.delete_credentials(vault_id=self.vault_id, url=url)
+
+    @override
+    def set_credit_card(self, **kwargs: Unpack[CreditCardDict]) -> None:
+        _ = self.vault_client.set_credit_card(self.vault_id, **kwargs)
+
+    @override
+    def get_credit_card(self) -> CreditCardDict:
+        return self.vault_client.get_credit_card(self.vault_id).credit_card
+
+    @override
+    def list_credentials(self) -> list[Credential]:
+        return self.vault_client.list_credentials(self.vault_id).credentials
+
+    @override
+    def delete_credit_card(self) -> None:
+        _ = self.vault_client.delete_credit_card(self.vault_id)
 
 
 @final
@@ -261,8 +307,6 @@ class VaultsClient(BaseClient):
         ]
 
     def get(self, vault_id: str) -> NotteVault:
-        from notte_sdk.vault import NotteVault
-
         """
         Get vault by id
 
@@ -275,8 +319,6 @@ class VaultsClient(BaseClient):
         return NotteVault(vault_id)
 
     def create(self, **data: Unpack[VaultCreateRequestDict]) -> NotteVault:
-        from notte_sdk.vault import NotteVault
-
         """
         Create vault
